+++ yggdrasil_decision_forests/port/python/ydf/dataset/io/dataset_io.py
@@ -103,7 +103,8 @@ def _unroll_column(
 
 def _unroll_dict(
     src: dataset_io_types.DictInputValues,
-    dont_unroll_columns: Optional[Sequence[str]] = None,
+    single_dim_columns: Sequence[str],
+    not_unrolled_multi_dim_columns: Sequence[str],
 ) -> Tuple[
     dataset_io_types.DictInputValues, dataset_io_types.UnrolledFeaturesInfo
 ]:
@@ -111,29 +112,33 @@ def _unroll_dict(
 
   Args:
     src: Dictionary of single and multi-dim values.
-    dont_unroll_columns: List of columns that cannot be unrolled. If one such
-      column needs to be unrolled, raise an error.
+    single_dim_columns: List of columns that should be single-dimensional. If
+      one such column is multi-dimensional, raise an error.
+    not_unrolled_multi_dim_columns: List of columns that should be
+      multi-dimensional and not unrolled.
 
   Returns:
     Dictionary containing only single-dimensional values.
   """
 
   # Index the columns for fast query.
-  dont_unroll_columns_set = (
-      set(dont_unroll_columns) if dont_unroll_columns else set()
-  )
+  single_dim_columns_set = set(single_dim_columns)
+  not_unrolled_multi_dim_columns_set = set(not_unrolled_multi_dim_columns)
 
   unrolled_features_info = {}
 
-  # Note: We only create a one dictionary independently of the number of
+  # Note: We only create a single dictionary independently of the number of
   # features.
   dst = {}
   for name, value in src.items():
+    if name in not_unrolled_multi_dim_columns_set:
+      dst[name] = value
+      continue
 
     any_is_unrolling = False
     sub_dst = {}
     for sub_name, sub_value, is_unrolling in _unroll_column(
-        name, value, allow_unroll=name not in dont_unroll_columns_set
+        name, value, allow_unroll=name not in single_dim_columns_set
     ):
       sub_dst[sub_name] = sub_value
       any_is_unrolling |= is_unrolling
@@ -148,7 +153,8 @@ def _unroll_dict(
 
 def cast_input_dataset_to_dict(
     data: dataset_io_types.IODataset,
-    dont_unroll_columns: Optional[Sequence[str]] = None,
+    single_dim_columns: Optional[Sequence[str]] = None,
+    not_unrolled_multi_dim_columns: Optional[Sequence[str]] = None,
 ) -> Tuple[
     dataset_io_types.DictInputValues, dataset_io_types.UnrolledFeaturesInfo
 ]:
@@ -158,14 +164,19 @@ def cast_input_dataset_to_dict(
 
   Args:
     data: Input data.
-    dont_unroll_columns: Column in "dont_unroll_columns" will not be unrolled.
+    single_dim_columns: Optional list of columns that should be
+      single-dimensional. If one such column is multi-dimensional, raise an
+      error.
+    not_unrolled_multi_dim_columns: Optional list of columns that should be
+      multi-dimensional and not unrolled.
 
   Returns:
     The normalized features, and information about unrolled features.
   """
 
   unroll_dict_kwargs = {
-      "dont_unroll_columns": dont_unroll_columns,
+      "single_dim_columns": single_dim_columns or [],
+      "not_unrolled_multi_dim_columns": not_unrolled_multi_dim_columns or [],
   }
 
   if pandas_io.is_pandas_dataframe(data):